+++ types/src/block/input/utxo.rs
@@ -8,7 +8,7 @@ use derive_more::From;
 use crate::block::{output::OutputId, payload::transaction::TransactionId, Error};
 
 /// Represents an input referencing an output.
-#[derive(Clone, Eq, PartialEq, Hash, Ord, PartialOrd, From, packable::Packable)]
+#[derive(Clone, Copy, Eq, PartialEq, Hash, Ord, PartialOrd, From, packable::Packable)]
 pub struct UtxoInput(OutputId);
 
 impl UtxoInput {