+++ types/tests/input/utxo.rs
@@ -21,6 +21,23 @@ fn kind() {
 }
 
 #[test]
+fn is_methods() {
+    let input = Input::from(UtxoInput::from_str(OUTPUT_ID).unwrap());
+
+    assert!(input.is_utxo());
+    assert!(!input.is_treasury());
+}
+
+#[test]
+fn as_methods() {
+    let utxo_input = UtxoInput::from_str(OUTPUT_ID).unwrap();
+    let input = Input::from(utxo_input);
+
+    assert_eq!(input.as_utxo(), &utxo_input);
+    assert!(std::panic::catch_unwind(|| input.as_treasury()).is_err());
+}
+
+#[test]
 fn debug_impl() {
     assert_eq!(
         format!("{:?}", UtxoInput::from_str(OUTPUT_ID).unwrap()),