+++ types/CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `OutputMetadata` moved from `iota-client`;
 - Derive `serde::{Serialize, Deserialize}` on `ChainId`;
 - `Input::{is_utxo, as_utxo, is_treasury, as_treasury}`;
+- Derive `Copy` on `UtxoInput`;
 
 ### Changed